Rename files  and folder to lowercase
