+++ keyboards/noxary/268_2_rgb/info.json
