+++ keyboards/noxary/268_2_rgb/keymaps/default/keymap.c
